Fix complete buttons not disappearing / Add ProxyProvider
